--- screens/SignInScreen.js
@@ -1,18 +1,0 @@
-import React, { Component } from "react";
-import { Text, View } from "react-native";
-
-export class SignInScreen extends Component {
-  render() {
-    return (
-      <View>
-        <Text> Sign In Screen! </Text>
-      </View>
-    );
-  }
-}
-
-SignInScreen.navigationOptions = {
-  title: "Sign In Screen"
-};
-
-export default SignInScreen;
